Added several updates, primarily for speed improvement.  Now the user can select whether (s)he wants to fit for the upper limit, which takes the bulk of the computational time.  For a bright enough source, this step is unnecessary, so lots of time is saved.  For further time saving, the user can choose to not fit for the FWHM of the image, which is an unnecessary step if the user specifies a pre-decided aperture size.  A conditional statement was added to do_dir_phot which will attempt to run photometry w/o the time consuming upper limit step first, and then return to find the upper limit should no target magnitude be found.
Finally, the S/N of the faintest target found by sextractor is saved in the output photometry dictionary under the keyword faintest_s2n.  The desired value of this parameter is set in pairitel_photo.sex; currently set to 3.0.  In practice, with enough stars in the photometered image, the average value of faintest_s2n should equal 3.0; however, on an image by image basis, the actual value can vary somewhat. 

git-svn-id: svn+ssh://lyra.berkeley.edu/Bloom/amorgan/SVNREPO/q_soft@245 15f48517-4a6d-0410-94cc-8a7d56af1ecd