Change in algorithm based on the description below:
What we're faced with is a two-step estimation problem.  1) going from features, X, to class probabilities, p-hat.  2) going from class probabilities to alpha-hat.  For the first step, we are using random forest.  For the second step, we are doing an empirical estimate, something like
alpha-hat_i = proportion of p-hat's smaller than p-hat_i

When we phrase the problem like this, I think it's more clear what the appropriate analysis is.

Right now, this is what (I think) we are doing:
Train a RF on the training set, compute p-hats for the set of GRBs w/o known z.  Map those p-hats to alpha-hats using the empirical proportion of training set p-hats smaller than each testing set p-hat.

This is what I think we should be doing:
Compute cross-validated p-hats for all of the training set (using RF).  From these p-hats, compute an empirical estimate of g, the function that maps from p-hats to alpha-hats.  Save this estimate, g-hat.  Now, train a random forest on all of the training data & predict p-hats for all of the data w/o known z.  Finally, use the cross-validated estimate of g-hat to map these p-hats to alpha-hats.

This feels right because: 1) we never want to compare in-sample to out-of-sample p-hat estimates, which is what is done in our current method.  Estimating g through cross-validation allows us to avoid this.  2) It seems like my proposal is closer to what we are actually doing during the model evaluation stage, where cross-validation is used to compute the p-hats and alpha-hats.