Updated qRSS to include the function "MonitorFile" - seemed like a reasonable place to put this, even though it's not explicitly an RSS related thing. Whatever, whatever, I do what I want!
This function periodically performes an action (eventually meant to download data from a website, but now just a print statement - perhaps can be expanded to perform any function).

At the same time, it tracks the status of a file (nominally latest_grb_time.txt) to see whether it has been updated. If it has, perform the aforementioned action faster, increasing this time periodically up to a max value.