added a self.covindex attribute to the param objects (so that we know the corresponding indices in the returned covariance matrix).
Added a  plot_marginalization() function, which, given a covariance matrix, and the indices of the two paramters you care about, and plot the contours corresponding to Delta_chisq of various values.

Also added a little wrapper function plot_marg_from_fitdict() which will take a fitdict, and a tuple of strings that you want to marginalize over from that fitdict.