Several tests and updates (see comments in code for full details)
In trying to verify that the least squares minimization technique is giving me the right solution, I implemented a different minimization using fmin and an explicit definition of the chi2 statistic i want to minimize.  This is a lot slower than the leastsq option and uses a bit of a dumber algorithm, but for low-dimensional problems they converge to the same solution (and the fmin seems to be more robust against initial starting values compared to leastsq for the test case i implemented).

However, for high dimensional problems, the fmin solution takes exceedingly long and is converging to a poor result (if at all), compared to leastsq.

Talking with John Rice - the Simplex algorithm that fmin uses (and fmin_bfgs, etc) are slow and inefficient for high dimensional problems; If they're taking long time to converge, and producing a worse chi-sq, he would not particularly worry that they are not giving the same or as good an answer as leastsq
Upshot  - I don't have a particular reason not to trust the leastsq solution.  As a sanity check, testing a bunch of different initial conditions over a range of sane values to see that leastsq converges to the same solution will be nice.